removed compiler warnings
git-svn-id: https://141.76.40.86/svn-students/iircommon/toolkit/trunk@1815 5654db93-19e5-4567-beb8-64e8c25c7c45

Former-commit-id: 5f7152b5be07f6c7ed7e69d2112dbefb0f3d0ef2

+++ src/test/java/ws/palladian/helper/CollectionHelperTest.java
@@ -21,7 +21,7 @@ public class CollectionHelperTest {
         Assert.assertEquals(2, (int) map.get(2.));
         Assert.assertEquals(2, (int) map.get(3.));
         Assert.assertEquals(1, (int) map.get(4.));
-        Assert.assertEquals(4, (int) map.size());
+        Assert.assertEquals(4, map.size());
     }
 
 }
